mcp-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
boto3==1.34.40
pydantic==2.6.1
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.1
jira==3.6.0
